# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Channel Doctor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /channels/search:
    get:
      operationId: searchChannels
      tags: [channels]
      summary: Search YouTube for channels by query (handle, URL, or name)
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        "200":
          description: Top results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelSearchResults"
...
  /connected/saved-analyses:
    get:
      operationId: listSavedAnalyses
      tags: [connected]
      summary: List previously run analyses for the signed-in user
      responses:
        "200":
          description: Saved analyses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedAnalysisList"
    post:
      operationId: createSavedAnalysis
      tags: [connected]
      summary: Run and save an analysis for a channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectChannelBody"
      responses:
        "200":
          description: Saved analysis created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedAnalysis"
  /connected/saved-analyses/{id}:
    delete:
      operationId: deleteSavedAnalysis
      tags: [connected]
      summary: Delete a saved analysis by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    ChannelSearchResult:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        thumbnail: { type: string }
        description: { type: string }
        customUrl: { type: string, nullable: true }
      required: [id, title, thumbnail, description]
    ChannelSearchResults:
      type: object
      properties:
        results:
          type: array
          items: { $ref: "#/components/schemas/ChannelSearchResult" }
      required: [results]
...
    ConnectChannelBody:
      type: object
      properties:
        channelId: { type: string, minLength: 1 }
      required: [channelId]
    SavedAnalysis:
      type: object
      properties:
        id: { type: string }
        channelId: { type: string }
        channelTitle: { type: string }
        channelThumbnail: { type: string }
        healthScore: { type: number }
        diagnosis: { type: string }
        createdAt: { type: string }
      required: [id, channelId, channelTitle, channelThumbnail, healthScore, diagnosis, createdAt]
    SavedAnalysisList:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/SavedAnalysis" }
      required: [items]
...